app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.19
pydantic==1.10.9
python-jose==3.3.0
python-multipart==0.0.6
aiofiles==23.1.0
Pillow==10.1.0
requests==2.31.0
bcrypt<4
passlib[bcrypt]
PyJWT==2.10.1
boto3==1.34.0
authlib==1.2.1
botocore==1.34.0

# AWS Dependencies (RDS & DynamoDB):
psycopg2-binary==2.9.7


